Minor refactor of sections service getSections function, moved files around and renamed
